+++ plugin-dev/Source/Sentry/Private/Android/SentrySamplingContextAndroid.h
@@ -13,7 +13,7 @@ public:
 
 	void SetupClassMethods();
 
-	virtual USentryTransactionContext* GetTransactionContext() const override;
+	virtual TSharedPtr<ISentryTransactionContext> GetTransactionContext() const override;
 	virtual TMap<FString, FString> GetCustomSamplingContext() const override;
 
 private: