+++ plugin-dev/Source/Sentry/Private/Apple/SentryScopeApple.h
@@ -15,9 +15,9 @@ public:
 
 	SentryScope* GetNativeObject();
 
-	virtual void AddBreadcrumb(USentryBreadcrumb* breadcrumb) override;
+	virtual void AddBreadcrumb(TSharedPtr<ISentryBreadcrumb> breadcrumb) override;
 	virtual void ClearBreadcrumbs() override;
-	virtual void AddAttachment(USentryAttachment* attachment) override;
+	virtual void AddAttachment(TSharedPtr<ISentryAttachment> attachment) override;
 	virtual void ClearAttachments() override;
 	virtual void SetTagValue(const FString& key, const FString& value) override;
 	virtual FString GetTagValue(const FString& key) const override;