+++ plugin-dev/Source/Sentry/Private/Desktop/SentryScopeDesktop.h
@@ -9,10 +9,6 @@
 
 #if USE_SENTRY_NATIVE
 
-class USentryEvent;
-class USentryBreadcrumb;
-class USentryAttachment;
-
 class SentryBreadcrumbDesktop;
 class SentryEventDesktop;
 
@@ -23,9 +19,9 @@ public:
 	SentryScopeDesktop(const SentryScopeDesktop& Scope);
 	virtual ~SentryScopeDesktop() override;
 
-	virtual void AddBreadcrumb(USentryBreadcrumb* breadcrumb) override;
+	virtual void AddBreadcrumb(TSharedPtr<ISentryBreadcrumb> breadcrumb) override;
 	virtual void ClearBreadcrumbs() override;
-	virtual void AddAttachment(USentryAttachment* attachment) override;
+	virtual void AddAttachment(TSharedPtr<ISentryAttachment> attachment) override;
 	virtual void ClearAttachments() override;
 	virtual void SetTagValue(const FString& key, const FString& value) override;
 	virtual FString GetTagValue(const FString& key) const override;
@@ -50,7 +46,6 @@ public:
 	virtual void Clear() override;
 
 	void Apply(TSharedPtr<SentryEventDesktop> event);
-	void AddBreadcrumb(TSharedPtr<SentryBreadcrumbDesktop> breadcrumb);
 
 private:
 	FString Dist;